+++ packages/falcon-server/src/containers/EndpointContainer.ts
@@ -1,10 +1,15 @@
 /* eslint-disable no-restricted-syntax, no-await-in-loop, no-underscore-dangle */
-import { EndpointEntry, EndpointConstructor } from '@deity/falcon-server-env';
+import { Cache, EndpointEntry, EndpointConstructor } from '@deity/falcon-server-env';
+import { EventEmitter2 } from 'eventemitter2';
 import { EndpointEntryMap } from '../types';
 import { BaseContainer } from './BaseContainer';
 
 export class EndpointContainer extends BaseContainer {
   public entries: Array<EndpointEntry> = [];
+
+  constructor(eventEmitter: EventEmitter2, protected cache: Cache) {
+    super(eventEmitter);
+  }
 
   /**
    * Instantiates endpoints based on the passed configuration and registers event handlers for them
@@ -21,6 +26,7 @@ export class EndpointContainer extends BaseContainer {
           return;
         }
         const endpointManager = new EndpointManagerClass({
+          cache: this.cache,
           config: endpointManagerConfig.config || {},
           name: endpointKey,
           eventEmitter: this.eventEmitter