+++ packages/falcon-shop-data/src/Customer/ResetCustomerPasswordMutation.tsx
@@ -3,7 +3,7 @@ import { useMutation, MutationHookOptions } from '@apollo/react-hooks';
 import { Mutation, OperationInput } from '@deity/falcon-data';
 import { ResetPasswordInput } from '@deity/falcon-shop-extension';
 
-export const RESET_PASSWORD_MUTATION = gql`
+export const RESET_PASSWORD = gql`
   mutation ResetPassword($input: ResetPasswordInput!) {
     resetPassword(input: $input)
   }
@@ -15,10 +15,10 @@ export type ResetPasswordResponse = {
 
 export class ResetPasswordMutation extends Mutation<ResetPasswordResponse, OperationInput<ResetPasswordInput>> {
   static defaultProps = {
-    mutation: RESET_PASSWORD_MUTATION
+    mutation: RESET_PASSWORD
   };
 }
 
 export const useResetPasswordMutation = (
   options: MutationHookOptions<ResetPasswordResponse, OperationInput<ResetPasswordInput>> = {}
-) => useMutation(RESET_PASSWORD_MUTATION, options);
+) => useMutation(RESET_PASSWORD, options);