+++ packages/falcon-ui/src/theme/index.ts
@@ -2,46 +2,37 @@ import CSS from 'csstype';
 import { Keyframes } from '@emotion/serialize';
 import { defaultBaseTheme } from './theme';
 import { PropsMappings } from './responsiveprops';
-import { mergeThemes } from './utils';
+import { mergeThemes, RecursivePartial } from './utils';
+
+// export themed component factory
+export * from './themed';
+export * from './utils';
+export * from './responsiveprops';
 
 export function createTheme(themeOverride: RecursivePartial<Theme> = {}): Theme {
   return mergeThemes(defaultBaseTheme, themeOverride);
 }
 
-// export themed component factory
-export * from './themed';
-
-export * from './utils';
-
-export * from './responsiveprops';
-
 // --- exported type definitions for theme  ----
 export interface Theme {
   colors: ThemeColors;
-  breakpoints: ThemeBreakpoints;
-  spacing: ThemeSpacing;
+  breakpoints: Readonly<ThemeBreakpoints>;
+  spacing: Readonly<ThemeSpacing>;
   fonts: ThemeFonts;
-  fontSizes: ThemeFontSizes;
+  fontSizes: Readonly<ThemeFontSizes>;
   fontWeights: ThemeFontWeights;
-  lineHeights: ThemeLineHeights;
-  letterSpacings: ThemeLetterSpacings;
+  lineHeights: Readonly<ThemeLineHeights>;
+  letterSpacings: Readonly<ThemeLetterSpacings>;
   borders: ThemeBorders;
-  borderRadius: ThemeBorderRadius;
+  borderRadius: Readonly<ThemeBorderRadius>;
   boxShadows: ThemeBoxShadows;
   easingFunctions: ThemeEasingFunctions;
   transitionDurations: ThemeTransitionDurations;
   keyframes: ThemeKeyframes;
   zIndex: ThemeZIndex;
-  components: ThemedComponents;
-  icons: ThemedIcons;
+  components: ThemeComponents;
+  icons: ThemeIcons;
 }
-type s = keyof Theme;
-
-export type ThemedIcons = {
-  [name: string]: {
-    icon: React.ComponentType | ((props: any) => JSX.Element);
-  } & ThemedComponentProps;
-};
 
 type ThemedPropMapping = {
   themeProp: keyof Theme;
@@ -52,8 +43,6 @@ type CssProps = CSS.PropertiesFallback<number | string>;
 type ResponsivePropMapping = {
   cssProp: keyof CssProps;
 };
-
-export type RecursivePartial<T> = { [key in keyof T]?: RecursivePartial<T[key]> };
 
 type CSSPseudoObject = {
   [key in CSS.SimplePseudos]?: CSSObject;
@@ -68,71 +57,74 @@ type CssResponsiveProps = {
 
 export interface CSSObject extends CssResponsiveProps, CSSPseudoObject, CSSOthersObject {}
 
-export interface PropsWithTheme {
-  theme: Theme;
-}
+export type PropsWithTheme<TProps> = TProps & { theme: Theme };
 
-export type BaseThemedComponentProps = {
-  [ComponentProp in keyof PropsMappings]?:
-    | (PropsMappings[ComponentProp] extends ThemedPropMapping
-        ? Extract<keyof Theme[PropsMappings[ComponentProp]['themeProp']], string>
-        : PropsMappings[ComponentProp] extends ResponsivePropMapping
-        ? CssProps[PropsMappings[ComponentProp]['cssProp']]
-        : (string | number))
-    | {
-        [Breakpoint in keyof Theme['breakpoints']]?: PropsMappings[ComponentProp] extends ThemedPropMapping
-          ? Extract<keyof Theme[PropsMappings[ComponentProp]['themeProp']], string>
-          : PropsMappings[ComponentProp] extends ResponsivePropMapping
-          ? CssProps[PropsMappings[ComponentProp]['cssProp']]
-          : (string | number);
-      };
+type ThemePropMap<TProp extends keyof PropsMappings> = PropsMappings[TProp] extends ThemedPropMapping
+  ? Extract<keyof Theme[PropsMappings[TProp]['themeProp']], string>
+  : PropsMappings[TProp] extends ResponsivePropMapping
+  ? CssProps[PropsMappings[TProp]['cssProp']]
+  : (string | number);
+
+export type BaseThemingProps = {
+  [TProp in keyof PropsMappings]?:
+    | ThemePropMap<TProp>
+    | { [TBreakpoint in keyof ThemeBreakpoints]?: ThemePropMap<TProp> };
 };
+export type InlineCss<TProps> = ((props: PropsWithTheme<TProps>) => CSSObject) | CSSObject;
 
-export type InlineCss<TProps> = ((props: PropsWithTheme & TProps) => CSSObject) | CSSObject;
-
-export interface ThemedComponentProps<TProps = {}> extends BaseThemedComponentProps {
+/** old `ThemedComponentProps` */
+export type ThemingProps<TProps = any> = BaseThemingProps & {
+  // as: Tag
+  // defaultTheme?: ComponentTheme<TProps> | { [name: string]: ComponentTheme<TProps> };
+  variant?: string;
   css?: InlineCss<TProps>;
-}
-
-export type ThemedComponentPropsWithVariants<TProps = {}> = ThemedComponentProps<TProps> & {
-  variants?: {
-    [variantKey: string]: ThemedComponentProps<TProps>;
-  };
 };
 
-export interface ThemedComponents {
-  [key: string]: ThemedComponentPropsWithVariants;
+export type PropsWithThemingProps<TProps> = TProps & ThemingProps<TProps>;
+
+type ComponentThemeVariant<TProps> = BaseThemingProps & { css?: InlineCss<TProps> };
+export type ComponentTheme<TProps> = ComponentThemeVariant<TProps> & {
+  variants?: { [variantKey: string]: ComponentThemeVariant<TProps> };
+};
+
+export interface ThemeComponents {
+  [key: string]: ComponentTheme<{}>;
 }
-type NumberOrStringValues<T> = { readonly [P in keyof T]: number | string };
+
+export type ThemeIcons = {
+  [name: string]: {
+    icon: React.ComponentType | ((props: any) => JSX.Element);
+  } & ThemingProps<any>;
+};
 
 type Colors = typeof defaultBaseTheme.colors;
 export interface ThemeColors extends Colors {}
 
-type Breakpoints = NumberOrStringValues<typeof defaultBaseTheme.breakpoints>;
+type Breakpoints = Record<keyof typeof defaultBaseTheme.breakpoints, number | string>;
 export interface ThemeBreakpoints extends Breakpoints {}
 
-type Spacing = NumberOrStringValues<typeof defaultBaseTheme.spacing>;
+type Spacing = Record<keyof typeof defaultBaseTheme.spacing, number | string>;
 export interface ThemeSpacing extends Spacing {}
 
 type Fonts = typeof defaultBaseTheme.fonts;
 export interface ThemeFonts extends Fonts {}
 
-type FontSizes = NumberOrStringValues<typeof defaultBaseTheme.fontSizes>;
+type FontSizes = Record<keyof typeof defaultBaseTheme.fontSizes, number | string>;
 export interface ThemeFontSizes extends FontSizes {}
 
 type FontWeights = typeof defaultBaseTheme.fontWeights;
 export interface ThemeFontWeights extends FontWeights {}
 
-type LineHeights = NumberOrStringValues<typeof defaultBaseTheme.lineHeights>;
+type LineHeights = Record<keyof typeof defaultBaseTheme.lineHeights, number | string>;
 export interface ThemeLineHeights extends LineHeights {}
 
-type LetterSpacings = NumberOrStringValues<typeof defaultBaseTheme.letterSpacings>;
+type LetterSpacings = Record<keyof typeof defaultBaseTheme.letterSpacings, number | string>;
 export interface ThemeLetterSpacings extends LetterSpacings {}
 
 type Borders = typeof defaultBaseTheme.borders;
 export interface ThemeBorders extends Borders {}
 
-type BorderRadius = NumberOrStringValues<typeof defaultBaseTheme.borderRadius>;
+type BorderRadius = Record<keyof typeof defaultBaseTheme.borderRadius, number | string>;
 export interface ThemeBorderRadius extends BorderRadius {}
 
 type BoxShadows = typeof defaultBaseTheme.boxShadows;