+++ CHANGELOG.md
@@ -21,12 +21,20 @@ Versions marked with a number and date (e.g. Falcon Client v0.1.0 (2018-10-05)) 
 - fixed handling of redirects by serviceworker ([#594](https://github.com/deity-io/falcon/pull/594))
 - fixed the `next` value for sign-in redirects ([#689](https://github.com/deity-io/falcon/pull/689))
 
+### Falcon I18n vNext
+
+- added `useI18n` hook ([#717](https://github.com/deity-io/falcon/pull/717))
+
 ### Falcon UI vNext
 
 - fixed theme merging ([#551](https://github.com/deity-io/falcon/pull/551))
 - fixed duplicated conditions when extracting themed css ([#611](https://github.com/deity-io/falcon/pull/611))
 - fixed incompatible `css` prop typings ([#645](https://github.com/deity-io/falcon/pull/645))
 - added `console.error` when an icon is not defined in the theme during development ([#682](https://github.com/deity-io/falcon/pull/682))
+- improved typings ([#669](https://github.com/deity-io/falcon/pull/669))
+  * renamed `ThemedComponentPropsWithVariants` to `ComponentTheme`
+  * renamed `ThemedComponentProps` to `ThemingProps`
+  * renamed `BaseThemedComponentProps` to `BaseThemingProps`
 
 ### Falcon E-commerce UI Kit discontinued
 
@@ -39,6 +47,7 @@ Versions marked with a number and date (e.g. Falcon Client v0.1.0 (2018-10-05)) 
 - fixed synchronization of account and sign in icon with customer logging in and out state ([#521](https://github.com/deity-io/falcon/pull/521))
 - added Apollo hooks support (upgrade `apollo-client` to version `2.6.x`) ([#596](https://github.com/deity-io/falcon/pull/596))
 - improved dynamic url resolution ([#618](https://github.com/deity-io/falcon/pull/618))
+- added `Add`/`EditAddressFormProvider` components to handle relevant form data ([#624](https://github.com/deity-io/falcon/pull/624))
 - added `Submit` component which provides context to form submit components ([#548](https://github.com/deity-io/falcon/pull/548))
 - fixed `LocaleSwitcher` which did not take into account `whitelist`ed languages, and `LocaleContext` which hardcoded language fallback instead of taking it from `clientConfig` ([#698](https://github.com/deity-io/falcon/pull/698))
 - extracted `CurrencyProvider` and `Currency` components from `LocaleProvider` and `Locale`, also introduced `useLocale` and `useCurrency` hooks ([#698](https://github.com/deity-io/falcon/pull/698))
@@ -51,6 +60,7 @@ Versions marked with a number and date (e.g. Falcon Client v0.1.0 (2018-10-05)) 
 - added Apollo hooks support (upgrade `apollo-client` to version `2.6.x`) ([#596](https://github.com/deity-io/falcon/pull/596))
 - fixed warning when `AddressDetails` receives non-unique values ([#614](https://github.com/deity-io/falcon/pull/614))
 - changed `Grid` component default `gridGap` to `md` ([#636](https://github.com/deity-io/falcon/pull/636))
+- added `AddressFormFields` and `Add`/`EditAddressForm` components to make address forms more reusable ([#624](https://github.com/deity-io/falcon/pull/624))
 - fixed `ProductCard` component which should not require `thumbnail` ([#685](https://github.com/deity-io/falcon/pull/685))
 - integrated `FormSubmit` with Formik, which allows it to hook into form state ([#548](https://github.com/deity-io/falcon/pull/548))
 - updated Formik to version 2.0 ([#705](https://github.com/deity-io/falcon/pull/705))
@@ -66,6 +76,7 @@ Versions marked with a number and date (e.g. Falcon Client v0.1.0 (2018-10-05)) 
 
 - initial release
 - added Apollo hooks support (upgrade `apollo-client` to version `2.6.x`) ([#596](https://github.com/deity-io/falcon/pull/596))
+- fixed response types of `AddressQuery` and `CountryListQuery` ([#624](https://github.com/deity-io/falcon/pull/624))
 
 ### Falcon Blog Data v1.0.0
 
@@ -89,6 +100,8 @@ Versions marked with a number and date (e.g. Falcon Client v0.1.0 (2018-10-05)) 
 - fixed priority for `ApiDataSource.getExtraResolvers` values ([#678](https://github.com/deity-io/falcon/pull/678))
 - `graphqlUtils` methods moved to `falcon-server-env` package ([#686](https://github.com/deity-io/falcon/pull/686))
 - simplified schema stitching call on Falcon-Server ([#701](https://github.com/deity-io/falcon/pull/701))
+- fixed `cache` exposing `EndpointContainer` ([#720](https://github.com/deity-io/falcon/pull/720))
+- exposed `dataSource` map for route context via `ctx.dataSources` ([#720](https://github.com/deity-io/falcon/pull/720))
 
 ### Falcon Server Env vNext
 
@@ -97,11 +110,17 @@ Versions marked with a number and date (e.g. Falcon Client v0.1.0 (2018-10-05)) 
 - added optional `ApiDataSource.getExtraResolvers` static method to define extra resolvers ([#557](https://github.com/deity-io/falcon/pull/557))
 - added re-export of all internally used components and helper utils ([#686](https://github.com/deity-io/falcon/pull/686))
 
+### Falcon Logger vNext
+
+- improved error handling for `logger-pretty` when working with incompatible log messages ([#716](https://github.com/deity-io/falcon/pull/716))
+
 ### Falcon Magento2 API vNext
 
 - added basic filtration into `productList` query resolver ([#575](https://github.com/deity-io/falcon/pull/575))
 - fixed `lastOrder` resolver ([#616](https://github.com/deity-io/falcon/pull/616))
 - improved dynamic url resolution ([#618](https://github.com/deity-io/falcon/pull/618))
+- changed return type of `PlaceOrder` mutation to `Order | PlaceOrder3dSecureResult` type ([#709](https://github.com/deity-io/falcon/pull/709))
+- configured TypeScript ([#468](https://github.com/deity-io/falcon/pull/468))
 
 ### Falcon Scripts vNext